requirements.txt
python-chess>=1.9.4
pandas>=1.5.0
numpy>=1.23.0
tqdm>=4.64.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
joblib>=1.2.0
xgboost>=1.7.0
flask>=2.3.0
torch>=2.0.0